requirements.txt
Django==5.0.7
django-environ==0.11.2
PyMySQL==1.1.1
django-ratelimit==4.1.0
django-csp==3.8
requests==2.32.3
huggingface-hub==0.23.5
gunicorn==22.0.0
pytest-django==4.8.0
django-extensions==3.2.3

